6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("Size of the Char: %lu bytes(s)\n", (unsigned long)sizeof(a));
-printf("Size of the int: %lu bytes(s)\n", (unsigned long)sizeof(b));
-printf("Size of the long: %lu bytes(s)\n", (unsigned long)sizeof(c));
-printf("Size of the long long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
-printf("Size of the float: %lu bytes(s)\n", (unsigned long)sizeof(f));
+printf("Size of the Char: %zu bytes(s)\n", sizeof(a));
+printf("Size of the int: %zu bytes(s)\n", sizeof(b));
+printf("Size of the long: %zu bytes(s)\n", sizeof(c));
+printf("Size of the long long int: %zu bytes(s)\n",sizeof(d));
+printf("Size of the float: %zu bytes(s)\n", sizeof(f));
 return (0);
 }
